requirements.txt
langchain==0.1.9
langchain-openai==0.0.7
langchainhub==0.1.14
faiss-cpu==1.7.4
python-dotenv==1.0.1